This reverts commit 28a23372185c7a30c7790085a6a010c3cad6ad09, reversing
changes made to e992f84735a2ca166a7388d4eca69a6b4e6ee212.

[[PR Description]]

Release Notes:

- (Added|Fixed|Improved) ...
([#<public_issue_number_if_exists>](https://github.com/zed-industries/community/issues/<public_issue_number_if_exists>)).

+++ crates/workspace/src/workspace.rs
@@ -4334,7 +4334,7 @@ impl Element for DisconnectedOverlay {
     }
 
     fn paint(&mut self, bounds: Bounds<Pixels>, overlay: &mut Self::State, cx: &mut WindowContext) {
-        cx.with_z_index(u16::MAX, |cx| {
+        cx.with_z_index(u8::MAX, |cx| {
             cx.add_opaque_layer(bounds);
             overlay.paint(cx);
         })
